refactor(compact_all_services): split helper into modules, add tests and runbook
- Replace monolithic helper.py with local_compaction, s3_compaction,
  migration, cleanup, and shared constants module.
- Update handler and delete_files_from_lookback_period imports; remove
  debug prints from scheduled compaction loop.
- Add unit tests and register services/compact_all_services/tests in CI.
- Document local vs manual S3 paths in READMEs and add operator runbook.
