+++ services/SERVICES_SUMMARY.md
@@ -8,7 +8,7 @@ Top-level folders under `services/` are summarized below. **Entryway** is the fi
 | `backfill` | Supports historical label generation: enqueueing posts missing ML labels, running integration runners against queues, writing cache buffers to storage, and separate PDS-oriented backfill helpers. | `pipelines/backfill_records_coordination/app.py` |
 | `calculate_analytics` | Operational tooling (for example per-integration record counts over date ranges) plus study-centric and one-off research analyses under `analyses/` and `study_analytics/`. | `services/calculate_analytics/count_records_for_integration.py` |
 | `calculate_superposters` | Computes users who post at unusually high volume so feeds can down-rank or otherwise treat superposters differently. | `pipelines/calculate_superposters/handler.py` |
-| `compact_all_services` | Walks configured service datasets and compacts storage (for example merging partition files) across the stack as part of storage hygiene. | `pipelines/compact_all_services/handler.py` |
+| `compact_all_services` | Walks configured service datasets and compacts storage (for example merging partition files) across the stack as part of storage hygiene. | `services/compact_all_services/local_compaction.py` (scheduled via `pipelines/compact_all_services/handler.py`) |
 | `compact_dedupe_data` | **Deprecated.** Older compaction and deduplication path backed by DynamoDB session tracking and Athena/Glue. | `pipelines/deprecated/compact_dedupe_data/handler.py` |
 | `compact_user_session_logs` | Compacts raw application session/interaction logs into forms suitable for downstream analytics. | `pipelines/compact_user_session_logs/handler.py` |
 | `consolidate_enrichment_integrations` | Merges preprocessed posts with outputs from ML integrations (labels, optional similarity scores) into consolidated enriched post records. | `pipelines/consolidate_enrichment_integrations/handler.py` |
