+++ services/compact_all_services/__init__.py
@@ -1,0 +1,1 @@
+"""Compact and reorganize study service datasets (local and optional S3 paths)."""
