+++ services/compact_all_services/constants.py
@@ -1,0 +1,5 @@
+"""Shared defaults for compact_all_services (no AWS clients)."""
+
+from typing import Literal
+
+default_export_format: Literal["json", "parquet"] = "parquet"
